Sync basetools' source and binary files with r1707 of the basetools project.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@9257 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/BaseTools/Source/C/Common/EfiUtilityMsgs.h
@@ -96,7 +96,7 @@ VOID
 DebugMsg (
   CHAR8   *FileName,
   UINT32  LineNumber,
-  UINT32  MsgLevel,
+  UINT64  MsgLevel,
   CHAR8   *OffendingText,
   CHAR8   *MsgFmt,
   ...
@@ -123,7 +123,7 @@ KeyMsg (
 
 VOID
 SetPrintLevel (
-  UINT32  LogLevel
+  UINT64  LogLevel
   );
 
 VOID
